first changes for multiple languages support
