+++ shop-app-materalize-frontend/src/app/[locale]/favicon.ico
