+++ shop-app-materalize-frontend/src/app/[locale]/globals.css
