+++ shop-app-materalize-frontend/src/app/[locale]/layout.tsx
