+++ shop-app-materalize-frontend/src/app/[locale]/main-page/page.tsx
@@ -3,6 +3,7 @@ import {Navbar} from "@/app/shared/Navbar";
 export default function MainPage(){
     return (<section id='MainPage' className='main-page-container'>
         Hello<br/>
+        <input type="checkbox" value="synthwave" className="toggle theme-controller"/>
 
         fsfasfd<br/>
