+++ shop-app-materalize-frontend/src/app/[locale]/manifest.json
