+++ shop-app-materalize-frontend/src/app/[locale]/page.tsx
