+++ shop-app-materalize-frontend/src/app/[locale]/register/page.tsx
