+++ shop-app-materalize-frontend/src/app/[locale]/shared/Footer.tsx
