--- shop-app-materalize-frontend/src/app/shared/Navbar.tsx
@@ -1,104 +1,0 @@
-"use client";
-import Link from 'next/link';
-import ThemeSwitcher from "@/app/context/ThemeSwitcher";
-import {ThemeContext} from "@/app/context/ThemeContext";
-import {useContext} from "react";
-
-export function Navbar(){
-    const { changeTheme } = useContext(ThemeContext);
-    return(<>
-        <div className="navbar bg-base-300 rounded-box shadow-xl">
-            <div className="navbar-start">
-                <div className="dropdown">
-                    <div tabIndex={0} role="button" className="btn btn-ghost lg:hidden">
-                        <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5" fill="none" viewBox="0 0 24 24"
-                             stroke="currentColor">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2"
-                                  d="M4 6h16M4 12h8m-8 6h16"/>
-                        </svg>
-                    </div>
-                    <ul tabIndex={1}
-                        className="menu menu-sm dropdown-content mt-3 z-[1] p-2 shadow bg-base-100 rounded-box w-52">
-                        <li><Link href="/" className='link'>Main page</Link></li>
-                        <li>
-                            <a>Parent</a>
-                            <ul className="p-2">
-                                <li><a>Submenu 1</a></li>
-                                <li><a>Submenu 2</a></li>
-                            </ul>
-                        </li>
-                        <li><a>Item 3</a></li>
-                    </ul>
-                </div>
-                <Link className="btn btn-ghost text-xl" href={"/"}>KKMK</Link>
-            </div>
-            <div className="navbar-center hidden lg:flex">
-                <ul className="menu menu-horizontal px-1">
-                    <li><Link href={"/"}>Item 1</Link></li>
-                    <li>
-                        <details>
-                            <summary>Parent</summary>
-                            <ul className="p-2">
-                                <li><a>Submenu 1</a></li>
-                                <li><a>Submenu 2</a></li>
-                            </ul>
-                        </details>
-                    </li>
-                    <li><a>Item 3</a></li>
-                </ul>
-            </div>
-
-            <div className="navbar-end">
-                <ThemeSwitcher handleOnClick={changeTheme}/>
-                <div className="flex-none">
-                    <ul className="menu menu-horizontal px-1">
-                        <li><Link href={"/login/"}>Login</Link></li>
-                        <li><Link href={"/register/"}>Register</Link></li>
-                    </ul>
-                </div>
-            </div>
-            {/*<div className="navbar-end">*/}
-            {/*    <div className="dropdown dropdown-end">*/}
-            {/*        <div tabIndex="0" role="button" className="btn btn-ghost btn-circle">*/}
-            {/*            <div className="indicator">*/}
-            {/*                <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5" fill="none" viewBox="0 0 24 24"*/}
-            {/*                     stroke="currentColor">*/}
-            {/*                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"*/}
-            {/*                          d="M3 3h2l.4 2M7 13h10l4-8H5.4M7 13L5.4 5M7 13l-2.293 2.293c-.63.63-.184 1.707.707 1.707H17m0 0a2 2 0 100 4 2 2 0 000-4zm-8 2a2 2 0 11-4 0 2 2 0 014 0z"/>*/}
-            {/*                </svg>*/}
-            {/*                <span className="badge badge-sm indicator-item">8</span>*/}
-            {/*            </div>*/}
-            {/*        </div>*/}
-            {/*        <div tabIndex="0" className="mt-3 z-[1] card card-compact dropdown-content w-52 bg-base-100 shadow">*/}
-            {/*            <div className="card-body">*/}
-            {/*                <span className="font-bold text-lg">8 Items</span>*/}
-            {/*                <span className="text-info">Subtotal: $999</span>*/}
-            {/*                <div className="card-actions">*/}
-            {/*                    <button className="btn btn-primary btn-block">View cart</button>*/}
-            {/*                </div>*/}
-            {/*            </div>*/}
-            {/*        </div>*/}
-            {/*    </div>*/}
-            {/*    <div class="dropdown dropdown-end">*/}
-            {/*        <div tabindex="0" role="button" class="btn btn-ghost btn-circle avatar">*/}
-            {/*            <div class="w-10 rounded-full">*/}
-            {/*                <img alt="Tailwind CSS Navbar component"*/}
-            {/*                     src="https://img.daisyui.com/images/stock/photo-1534528741775-53994a69daeb.jpg"/>*/}
-            {/*            </div>*/}
-            {/*        </div>*/}
-            {/*        <ul tabindex="0"*/}
-            {/*            class="menu menu-sm dropdown-content mt-3 z-[1] p-2 shadow bg-base-100 rounded-box w-52">*/}
-            {/*            <li>*/}
-            {/*                <a class="justify-between">*/}
-            {/*                    Profile*/}
-            {/*                    <span class="badge">New</span>*/}
-            {/*                </a>*/}
-            {/*            </li>*/}
-            {/*            <li><a>Settings</a></li>*/}
-            {/*            <li><a>Logout</a></li>*/}
-            {/*        </ul>*/}
-            {/*    </div>*/}
-            {/*</div>*/}
-        </div>
-    </>)
-}
